✨ Added `get_spawnable_species` function
This commit adds a new function `check_species_can_spawn` to the `exobiology` module. This function is used to check if a given species can spawn on a planet based on the planet scan event. It iterates over the spawn conditions of each species and returns `true` if all conditions are met.

The function `get_spawnable_species` has also been implemented, which returns a list of species that can spawn on the given planet.