Change Variant creation; Confirm Species name
Creating variants from a str only had a special case for `$Codex_Ent_Ground_Struct_Ice_Name;`.
Now it first checks if a species exists with the variant name. If it does we assume it doesn't have real variants and create a 'simple' variant.
Otherwise we create the 'complex' version of a variant.